# docs/swagger.yaml
basePath: /
definitions:
...
  ErrorEnvelope:
    properties:
      error:
        $ref: '#/definitions/ErrorDetails'
      requestId:
        example: c56a4180-65aa-42ec-a945-5fd21dec0538
        type: string
      success:
        example: false
        type: boolean
    required:
    - error
    - success
    type: object
...
  OrderResponse:
    properties:
      created_at:
        format: date-time
        type: string
      order_id:
        example: d0000000-0000-0000-0000-000000000001
        format: uuid
        type: string
      product_id:
        example: c0000000-0000-0000-0000-000000000001
        format: uuid
        type: string
      product_model:
        example: MacBook Pro 16 M3
        type: string
      quantity:
        example: 1
        minimum: 1
        type: integer
      remaining_balance:
        example: 500
        format: double
        minimum: 0
        type: number
      status:
        enum:
        - CREATED
        - PROCESSING
        - SHIPPED
        - DELIVERED
        - CANCELLED
        example: CREATED
        type: string
      total_price:
        example: 4998
        format: double
        minimum: 0
        type: number
      unit_price:
        example: 2499
        format: double
        minimum: 0
        type: number
      user_id:
        example: b0000000-0000-0000-0000-000000000002
        format: uuid
        type: string
    required:
    - created_at
    - order_id
    - product_id
    - product_model
    - quantity
    - status
    - total_price
    - unit_price
    type: object
...
  SuccessEnvelope:
    properties:
      data: {}
      page:
        example: 1
        type: integer
      page_size:
        example: 20
        type: integer
      pagination:
        $ref: '#/definitions/PaginationMetadata'
      requestId:
        example: c56a4180-65aa-42ec-a945-5fd21dec0538
        type: string
      success:
        example: true
        type: boolean
      total_count:
        example: 45
        type: integer
      total_pages:
        example: 3
        type: integer
    required:
    - data
    - success
    type: object
...
  UpdateOrderStatusRequest:
    properties:
      status:
        enum:
        - CREATED
        - PROCESSING
        - SHIPPED
        - DELIVERED
        - CANCELLED
        example: SHIPPED
        type: string
    required:
    - status
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Clean, lightweight, and idiomatic Go backend for warehouse inventory
    and transactional ordering.
  title: Warehouse REST API Testbench
  version: "1.0"
paths:
...
  /api/admin/orders:
    get:
      description: Retrieve all orders across the system for administrative auditing
        and fulfillment
      operationId: listAdminOrders
      produces:
      - application/json
      responses:
        "200":
          description: List of all orders
          schema:
            allOf:
            - $ref: '#/definitions/SuccessEnvelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/OrderResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "503":
          description: Service unavailable
          schema:
            $ref: '#/definitions/ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: List all orders (Admin)
      tags:
      - Admin
  /api/admin/orders/{id}/status:
    patch:
      consumes:
      - application/json
      description: Update the fulfillment status of an order (e.g. PROCESSING, SHIPPED,
        DELIVERED, CANCELLED)
      operationId: updateOrderStatus
      parameters:
      - description: Order ID (UUID)
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Order status payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/UpdateOrderStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Order status updated
          schema:
            allOf:
            - $ref: '#/definitions/SuccessEnvelope'
            - properties:
                data:
                  $ref: '#/definitions/OrderResponse'
              type: object
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "404":
          description: Order not found
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "422":
          description: Invalid status transition
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "503":
          description: Service unavailable
          schema:
            $ref: '#/definitions/ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Update order fulfillment status
      tags:
      - Admin
  /api/admin/products:
    get:
      description: Retrieve full unrestricted product inventory for administrators
      operationId: listAdminProducts
      produces:
      - application/json
      responses:
        "200":
          description: List of all products
          schema:
            allOf:
            - $ref: '#/definitions/SuccessEnvelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/Product'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "503":
          description: Service unavailable
          schema:
            $ref: '#/definitions/ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: List all warehouse products (Admin)
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: Add a new product item with arbitrary category, manufacturer, model,
        price, and stock
      operationId: createProduct
      parameters:
      - description: Product details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Product created
          schema:
            allOf:
            - $ref: '#/definitions/SuccessEnvelope'
            - properties:
                data:
                  $ref: '#/definitions/Product'
              type: object
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "503":
          description: Service unavailable
          schema:
            $ref: '#/definitions/ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Add a new product to warehouse
      tags:
      - Admin
  /api/admin/products/{id}/stock:
    patch:
      consumes:
      - application/json
      description: Set stock quantity for a warehouse product
      operationId: updateProductStock
      parameters:
      - description: Product ID (UUID)
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Stock update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/UpdateStockRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Stock updated
          schema:
            allOf:
            - $ref: '#/definitions/SuccessEnvelope'
            - properties:
                data:
                  $ref: '#/definitions/Product'
              type: object
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "503":
          description: Service unavailable
          schema:
            $ref: '#/definitions/ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Update product stock quantity
      tags:
      - Admin
  /api/admin/users:
    get:
      description: Query paginated list of users with optional role filtering
      operationId: listAdminUsers
      parameters:
      - description: Filter by role (admin or user)
        in: query
        name: role
        type: string
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 20, max 100)
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of users
          schema:
            allOf:
            - $ref: '#/definitions/SuccessEnvelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/UserSummary'
                  type: array
              type: object
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "503":
          description: Service unavailable
          schema:
            $ref: '#/definitions/ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: Create a new admin or regular user account
      operationId: createAdminUser
      parameters:
      - description: User details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User created
          schema:
            allOf:
            - $ref: '#/definitions/SuccessEnvelope'
            - properties:
                data:
                  $ref: '#/definitions/UserSummary'
              type: object
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "409":
          description: Username already exists
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "503":
          description: Service unavailable
          schema:
            $ref: '#/definitions/ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Create a user account
      tags:
      - Admin
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and JWT token. Example: "Bearer
      ey..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"